requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4